+++ dissociated-ipc/CMakeLists.txt
@@ -86,7 +86,7 @@ target_link_libraries(ucx::ucx INTERFACE ucx::ucp ucx::uct ucx::ucs)
 add_definitions(-DFMT_USE_NONTYPE_TEMPLATE_ARGS=0)
 add_executable(
   arrow-cudf-flight
-  cudf-flight-poc.cc
+  cudf-flight-ucx.cc
   cudf-flight-client.cc
   cudf-flight-server.cc
   ucx_utils.cc
