+++ dissociated-ipc/cudf-flight-server.cc
@@ -27,7 +27,7 @@
 #include <arrow/util/logging.h>
 #include <arrow/util/uri.h>
 
-#include "cudf-flight-poc.h"
+#include "cudf-flight-ucx.h"
 #include "ucx_server.h"
 
 namespace flight = arrow::flight;
